+++ NEWS.md
@@ -1,5 +1,8 @@
 # bundle (development version)
 
+* Added bundle method for objects from `dbarts::bart()` and, by extension,
+  `parsnip::bart(engine = "dbarts")` (#64).
+
 # bundle 0.1.1
 
 * Fixed bundling of recipes steps situated inside of workflows.